refactor: split data/web/http_client.py into focused modules
